+++ .gitignore
@@ -4,3 +4,6 @@
 .idea
 .idea/
 .idea/*
+.vscode
+nw
+php-8.0.6
