+++ Noise.h
@@ -1,0 +1,4 @@
+#ifndef NOISELIB_NOISE_H
+#define NOISELIB_NOISE_H
+#include "Noise.cpp"
+#endif //NOISELIB_NOISE_H
