+++ php_noiselib.h
@@ -1,0 +1,10 @@
+#ifndef PHP_NOISELIB_NOISE_H
+#define PHP_NOISELIB_NOISE_H
+
+#include "php.h"
+
+extern zend_module_entry noiselib_module_entry;
+#define phpext_noise_ptr &noiselib_module_entry
+#define PHP_NOISELIB_VERSION "1.0.0"
+#endif
+
